+++ contracts/socket/andromeda-socket-astroport/examples/schema.rs
@@ -1,0 +1,12 @@
+use andromeda_socket::astroport::{ExecuteMsg, InstantiateMsg, QueryMsg};
+
+use cosmwasm_schema::write_api;
+
+fn main() {
+    write_api! {
+        instantiate: InstantiateMsg,
+        query: QueryMsg,
+        execute: ExecuteMsg,
+
+    }
+}